I update the GUI.
There are 1 main class that controls the other 3 subclasses

Start() --> The start menu, it asks the user to start or quit
USer() --> The game buttons show up and submit button
Wait() --> The waiting screen shows up waiting the other player to make a move.

The Wait() function has an extra temp button for now. It shouldn't stay there was the project is done.